Add pango to conda environment
Closes https://github.com/greenelab/manubot-rootstock/issues/144
Merges https://github.com/greenelab/manubot-rootstock/pull/146

pango is required by WeasyPrint during the PDF build. While usually
built-in to the OS on Linux, it is not pre-installed on macOS.
macOS were installing pango using homebrew, which led to a
segmentation fault error.

Pango is available on conda-forge for linux and mac:
https://anaconda.org/conda-forge/pango
https://github.com/conda-forge/pango-feedstock

Adding pango to environment.yml prevents conda environment
installation on Windows, but has the benefits of standardizing
the pango version on linux and mac platforms and preventing the
mac homebrew issue.